fix: `defineI18nRoute` macro transformed inside `<template>`
