+++ src/transform/macros.ts
@@ -7,11 +7,10 @@
  */
 
 import createDebug from 'debug'
-import { pathToFileURL } from 'node:url'
-import { createUnplugin } from 'unplugin'
-import { parseQuery, parseURL } from 'ufo'
 import MagicString from 'magic-string'
-import { VIRTUAL_PREFIX_HEX } from './utils'
+import { createUnplugin } from 'unplugin'
+import { parse as parseSFC } from '@vue/compiler-sfc'
+import { VIRTUAL_PREFIX_HEX, isVue } from './utils'
 import { NUXT_I18N_COMPOSABLE_DEFINE_ROUTE } from '../constants'
 
 export interface TransformMacroPluginOptions {
@@ -29,46 +28,48 @@ const debug = createDebug('@nuxtjs/i18n:transform:macros')
 export const TransformMacroPlugin = createUnplugin((options: TransformMacroPluginOptions) => {
   return {
     name: 'nuxtjs:i18n-macros-transform',
-    enforce: 'post',
+    enforce: 'pre',
 
     transformInclude(id) {
       if (!id || id.startsWith(VIRTUAL_PREFIX_HEX)) {
         return false
       }
-      const { pathname, search } = parseURL(decodeURIComponent(pathToFileURL(id).href))
-      return pathname.endsWith('.vue') || !!parseQuery(search).macro
+
+      return isVue(id, { type: ['script'] })
     },
 
     transform(code, id) {
       debug('transform', id)
 
-      const s = new MagicString(code)
-      const { search } = parseURL(decodeURIComponent(pathToFileURL(id).href))
+      const parsed = parseSFC(code, { sourceMap: false })
+      // only transform <script>
+      const script = parsed.descriptor.scriptSetup ?? parsed.descriptor.script
+      if (!script) {
+        return
+      }
 
-      function result() {
-        if (s.hasChanged()) {
-          debug('transformed: id -> ', id)
-          debug('transformed: code -> ', s.toString())
-          return {
-            code: s.toString(),
-            map: options.sourcemap ? s.generateMap({ hires: true }) : undefined
-          }
+      const s = new MagicString(code)
+
+      // match content inside <script>
+      const match = script.content.match(new RegExp(`\\b${NUXT_I18N_COMPOSABLE_DEFINE_ROUTE}\\s*\\(\\s*`))
+      if (match?.[0]) {
+        // tree-shake out any runtime references to the macro.
+        const scriptString = new MagicString(script.content)
+        scriptString.overwrite(match.index!, match.index! + match[0].length, `false && /*#__PURE__*/ ${match[0]}`)
+
+        // using the locations from the parsed result we only replace the <script> contents
+        s.overwrite(script.loc.start.offset, script.loc.end.offset, scriptString.toString())
+      }
+
+      if (s.hasChanged()) {
+        debug('transformed: id -> ', id)
+        debug('transformed: code -> ', s.toString())
+
+        return {
+          code: s.toString(),
+          map: options.sourcemap ? s.generateMap({ hires: true }) : undefined
         }
       }
-
-      // tree-shake out any runtime references to the macro.
-      // we do this first as it applies to all files, not just those with the query
-      const match = code.match(new RegExp(`\\b${NUXT_I18N_COMPOSABLE_DEFINE_ROUTE}\\s*\\(\\s*`))
-      if (match?.[0]) {
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        s.overwrite(match.index!, match.index! + match[0].length, `false && /*#__PURE__*/ ${match[0]}`)
-      }
-
-      if (!parseQuery(search).macro) {
-        return result()
-      }
-
-      return result()
     }
   }
 })
